+++ api/routes/habits.js
@@ -1,10 +1,11 @@
 const router = require("express").Router();
 const Habits = require("../models/Habits.js");
 
-router.get("/", async (req, res) => {
+router.get("/:owner", async (req, res) => {
+  const { owner } = req.params;
   try {
-    const goodhab = await Habits.find({ typeofhabit: "good" });
-    const badhab = await Habits.find({ typeofhabit: "bad" });
+    const goodhab = await Habits.find({ typeofhabit: "good", user: owner });
+    const badhab = await Habits.find({ typeofhabit: "bad", user: owner });
     res.send({ ok: true, good: goodhab, bad: badhab });
   } catch (error) {
     res.send({ ok: false, message: error });
@@ -12,11 +13,16 @@ router.get("/", async (req, res) => {
 });
 
 router.post("/newhabit", async (req, res) => {
-  const { name, typeofhabit, color } = req.body;
+  const { name, typeofhabit, color, owner } = req.body;
   //console.log(name);
   //console.log(typeofhabit);
   try {
-    const newHabit = { name: name, typeofhabit: typeofhabit, color: color };
+    const newHabit = {
+      name: name,
+      typeofhabit: typeofhabit,
+      color: color,
+      user: owner,
+    };
     const hab = await Habits.create(newHabit);
     //console.log(hab);
     res.send({ ok: true, data: hab });
@@ -26,11 +32,11 @@ router.post("/newhabit", async (req, res) => {
 });
 
 router.patch("/updatehabit", async (req, res) => {
-  const { name, typeofhabit, color } = req.body;
+  const { name, typeofhabit, color, owner } = req.body;
   try {
     const hab = await Habits.findOneAndUpdate(
       { name: name },
-      { name: name, typeofhabit: typeofhabit, color: color },
+      { name: name, typeofhabit: typeofhabit, color: color, user: owner },
     );
     res.send({ ok: true, data: hab });
   } catch (error) {
@@ -38,10 +44,10 @@ router.patch("/updatehabit", async (req, res) => {
   }
 });
 
-router.delete("/deletehabit/:hab", async (req, res) => {
-  const { hab } = req.params;
+router.delete("/deletehabit/:hab/:owner", async (req, res) => {
+  const { hab, owner } = req.params;
   try {
-    const delhab = await Habits.findOneAndDelete({ name: hab });
+    const delhab = await Habits.findOneAndDelete({ name: hab, user: owner });
     res.send({ ok: true, data: delhab });
   } catch (error) {
     res.send({ ok: false, message: error });
